Navlist file created
